Conflicts:
	src/server/game/Achievements/AchievementMgr.cpp
	src/server/game/Battlegrounds/BattlegroundMgr.cpp
	src/server/game/DataStores/DBCStructure.h
	src/server/game/DataStores/DBCfmt.h
	src/server/game/Server/Protocol/Opcodes.cpp
	src/server/game/Server/Protocol/Opcodes.h

+++ src/server/game/Server/Protocol/Handlers/BattleGroundHandler.cpp
@@ -342,7 +342,7 @@ void WorldSession::HandleBattlefieldListOpcode(WorldPacket &recv_data)
 {
     sLog.outDebug("WORLD: Recvd CMSG_BATTLEFIELD_LIST Message");
 
-    uint32 bgTypeId;
+    uint8 bgTypeId;
     recv_data >> bgTypeId;                                  // id from DBC
 
     WorldPacket data;
